+++ src/Eloquent/Builder.php
@@ -326,12 +326,12 @@ class Builder extends EloquentBuilder
             if ($relation instanceof EmbedsOneOrMany) {
                 switch ($function) {
                     case 'count':
-                        $this->project([$alias => ['$size' => ['$ifNull' => ['$' . $name, []]]]]);
+                        $this->getQuery()->project([$alias => ['$size' => ['$ifNull' => ['$' . $name, []]]]]);
                         break;
                     case 'min':
                     case 'max':
                     case 'avg':
-                        $this->project([$alias => ['$' . $function => '$' . $name . '.' . $column]]);
+                        $this->getQuery()->project([$alias => ['$' . $function => '$' . $name . '.' . $column]]);
                         break;
                     default:
                         throw new InvalidArgumentException(sprintf('Invalid aggregate function "%s"', $function));
